requirements.txt
# 核心依赖
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 数据下载
yfinance>=0.2.28

# 技术指标
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# 深度学习
torch>=2.0.0
torchvision>=0.15.0

# 数据存储
pyarrow>=12.0.0
tables>=3.8.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 日志和实验跟踪
tqdm>=4.65.0
wandb>=0.15.0

# 工具
scikit-learn>=1.3.0
